ezyZip/requirements.txt
pyyaml
numpy
scikit-learn
matplotlib
reportlab
transformers
torch
huggingface-hub